Refactors all server ZMQ responses into handlers
