+++ src/bedrock/kvs/server.cpp
@@ -15,10 +15,10 @@
 #include "spdlog/spdlog.h"
 #include "hash_ring.hpp"
 #include "kvs/rc_pair_lattice.hpp"
-#include "kvs/kvs_handlers.hpp"
 #include "communication.pb.h"
 #include "zmq/socket_cache.hpp"
 #include "zmq/zmq_util.hpp"
+#include "kvs/kvs_handlers.hpp"
 #include "common.hpp"
 #include "yaml-cpp/yaml.h"
 #include "yaml-cpp/node/node.h"
@@ -33,8 +33,6 @@ const unsigned KEY_MONITORING_THRESHOLD = 60;
 const unsigned RETRY_THRESHOLD = 10;
 
 unsigned SELF_TIER_ID;
-
-// number of worker threads
 unsigned THREAD_NUM;
 
 unsigned MEMORY_THREAD_NUM;
@@ -47,244 +45,6 @@ unsigned DEFAULT_LOCAL_REPLICATION;
 // read-only per-tier metadata
 unordered_map<unsigned, TierData> tier_data_map;
 
-pair<ReadCommittedPairLattice<string>, unsigned> process_get(const string& key, Serializer* serializer) {
-  unsigned err_number = 0;
-  auto res = serializer->get(key, err_number);
-
-  // check if the value is an empty string
-  if (res.reveal().value == "") {
-    err_number = 1;
-  }
-  return pair<ReadCommittedPairLattice<string>, unsigned>(res, err_number);
-}
-
-void process_put(const string& key,
-    const unsigned long long& timestamp,
-    const string& value,
-    Serializer* serializer,
-    unordered_map<string, KeyStat>& key_stat_map) {
-
-  if (serializer->put(key, value, timestamp)) {
-    // update value size if the value is replaced
-    key_stat_map[key].size_ = value.size();
-  }
-}
-
-communication::Response process_request(
-    communication::Request& req,
-    unordered_set<string>& local_changeset,
-    Serializer* serializer,
-    ServerThread& wt,
-    unordered_map<unsigned, GlobalHashRing>& global_hash_ring_map,
-    unordered_map<unsigned, LocalHashRing>& local_hash_ring_map,
-    unordered_map<string, KeyInfo>& placement,
-    SocketCache& pushers,
-    unordered_map<string, KeyStat>& key_stat_map,
-    unordered_map<string, multiset<std::chrono::time_point<std::chrono::system_clock>>>& key_access_timestamp,
-    unsigned& total_access,
-    chrono::system_clock::time_point& start_time,
-    unordered_map<string, pair<chrono::system_clock::time_point, vector<PendingRequest>>>& pending_request_map,
-    unsigned& seed) {
-
-  communication::Response response;
-  string respond_id = "";
-
-  if (req.has_request_id()) {
-    respond_id = req.request_id();
-    response.set_response_id(respond_id);
-  }
-
-  vector<unsigned> tier_ids;
-  tier_ids.push_back(SELF_TIER_ID);
-  bool succeed;
-
-  if (req.type() == "GET") {
-    for (int i = 0; i < req.tuple_size(); i++) {
-      // first check if the thread is responsible for the key
-      string key = req.tuple(i).key();
-      auto threads = get_responsible_threads(wt.get_replication_factor_connect_addr(), key, is_metadata(key), global_hash_ring_map, local_hash_ring_map, placement, pushers, tier_ids, succeed, seed);
-
-      if (succeed) {
-        if (threads.find(wt) == threads.end()) {
-          if (is_metadata(key)) {
-            communication::Response_Tuple* tp = response.add_tuple();
-
-            tp->set_key(key);
-            tp->set_err_number(2);
-          } else {
-            //placement.erase(key);
-            issue_replication_factor_request(wt.get_replication_factor_connect_addr(), key, global_hash_ring_map[1], local_hash_ring_map[1], pushers, seed);
-            string val = "";
-
-            if (pending_request_map.find(key) == pending_request_map.end()) {
-              pending_request_map[key].first = chrono::system_clock::now();
-            }
-
-            pending_request_map[key].second.push_back(PendingRequest("G", val, req.respond_address(), respond_id));
-          }
-        } else {
-          communication::Response_Tuple* tp = response.add_tuple();
-          tp->set_key(key);
-
-          auto res = process_get(key, serializer);
-          tp->set_value(res.first.reveal().value);
-          tp->set_err_number(res.second);
-
-          if (req.tuple(i).has_num_address() && req.tuple(i).num_address() != threads.size()) {
-            tp->set_invalidate(true);
-          }
-
-          key_access_timestamp[key].insert(std::chrono::system_clock::now());
-          total_access += 1;
-        }
-      } else {
-        string val = "";
-
-        if (pending_request_map.find(key) == pending_request_map.end()) {
-          pending_request_map[key].first = chrono::system_clock::now();
-        }
-
-        pending_request_map[key].second.push_back(PendingRequest("G", val, req.respond_address(), respond_id));
-      }
-    }
-  } else if (req.type() == "PUT") {
-    for (int i = 0; i < req.tuple_size(); i++) {
-      // first check if the thread is responsible for the key
-      string key = req.tuple(i).key();
-      auto threads = get_responsible_threads(wt.get_replication_factor_connect_addr(), key, is_metadata(key), global_hash_ring_map, local_hash_ring_map, placement, pushers, tier_ids, succeed, seed);
-
-      if (succeed) {
-        if (threads.find(wt) == threads.end()) {
-          if (is_metadata(key)) {
-            communication::Response_Tuple* tp = response.add_tuple();
-
-            tp->set_key(key);
-            tp->set_err_number(2);
-          } else {
-            issue_replication_factor_request(wt.get_replication_factor_connect_addr(), key, global_hash_ring_map[1], local_hash_ring_map[1], pushers, seed);
-
-            if (pending_request_map.find(key) == pending_request_map.end()) {
-              pending_request_map[key].first = chrono::system_clock::now();
-            }
-
-            if (req.has_respond_address()) {
-              pending_request_map[key].second.push_back(PendingRequest("P", req.tuple(i).value(), req.respond_address(), respond_id));
-            } else {
-              pending_request_map[key].second.push_back(PendingRequest("P", req.tuple(i).value(), "", respond_id));
-            }
-          }
-        } else {
-          communication::Response_Tuple* tp = response.add_tuple();
-          tp->set_key(key);
-
-          auto current_time = chrono::system_clock::now();
-          auto ts = generate_timestamp(chrono::duration_cast<chrono::milliseconds>(current_time-start_time).count(), wt.get_tid());
-
-          process_put(key, ts, req.tuple(i).value(), serializer, key_stat_map);
-          tp->set_err_number(0);
-
-          if (req.tuple(i).has_num_address() && req.tuple(i).num_address() != threads.size()) {
-            tp->set_invalidate(true);
-          }
-
-          key_access_timestamp[key].insert(std::chrono::system_clock::now());
-          total_access += 1;
-          local_changeset.insert(key);
-        }
-      } else {
-        if (pending_request_map.find(key) == pending_request_map.end()) {
-          pending_request_map[key].first = chrono::system_clock::now();
-        }
-
-        if (req.has_respond_address()) {
-          pending_request_map[key].second.push_back(PendingRequest("P", req.tuple(i).value(), req.respond_address(), respond_id));
-        } else {
-          pending_request_map[key].second.push_back(PendingRequest("P", req.tuple(i).value(), "", respond_id));
-        }
-      }
-    }
-  }
-  return response;
-}
-
-void process_gossip(
-    communication::Request& gossip,
-    ServerThread& wt,
-    unordered_map<unsigned, GlobalHashRing>& global_hash_ring_map,
-    unordered_map<unsigned, LocalHashRing>& local_hash_ring_map,
-    unordered_map<string, KeyInfo>& placement,
-    SocketCache& pushers,
-    Serializer* serializer,
-    unordered_map<string, KeyStat>& key_stat_map,
-    unordered_map<string, pair<chrono::system_clock::time_point, vector<PendingGossip>>>& pending_gossip_map,
-    unsigned& seed) {
-  vector<unsigned> tier_ids;
-  tier_ids.push_back(SELF_TIER_ID);
-  bool succeed;
-  unordered_map<string, communication::Request> gossip_map;
-
-  for (int i = 0; i < gossip.tuple_size(); i++) {
-    // first check if the thread is responsible for the key
-    string key = gossip.tuple(i).key();
-    auto threads = get_responsible_threads(wt.get_replication_factor_connect_addr(), key, is_metadata(key), global_hash_ring_map, local_hash_ring_map, placement, pushers, tier_ids, succeed, seed);
-
-    if (succeed) {
-      if (threads.find(wt) != threads.end()) {
-        process_put(gossip.tuple(i).key(), gossip.tuple(i).timestamp(), gossip.tuple(i).value(), serializer, key_stat_map);
-      } else {
-        if (is_metadata(key)) {
-        // forward the gossip
-          for (auto it = threads.begin(); it != threads.end(); it++) {
-            if (gossip_map.find(it->get_gossip_connect_addr()) == gossip_map.end()) {
-              gossip_map[it->get_gossip_connect_addr()].set_type("PUT");
-            }
-
-            prepare_put_tuple(gossip_map[it->get_gossip_connect_addr()], key, gossip.tuple(i).value(), gossip.tuple(i).timestamp());
-          }
-        } else {
-          issue_replication_factor_request(wt.get_replication_factor_connect_addr(), key, global_hash_ring_map[1], local_hash_ring_map[1], pushers, seed);
-
-          if (pending_gossip_map.find(key) == pending_gossip_map.end()) {
-            pending_gossip_map[key].first = chrono::system_clock::now();
-          }
-
-          pending_gossip_map[key].second.push_back(PendingGossip(gossip.tuple(i).value(), gossip.tuple(i).timestamp()));
-        }
-      }
-    } else {
-      if (pending_gossip_map.find(key) == pending_gossip_map.end()) {
-        pending_gossip_map[key].first = chrono::system_clock::now();
-      }
-
-      pending_gossip_map[key].second.push_back(PendingGossip(gossip.tuple(i).value(), gossip.tuple(i).timestamp()));
-    }
-  }
-
-  // redirect gossip
-  for (auto it = gossip_map.begin(); it != gossip_map.end(); it++) {
-    push_request(it->second, pushers[it->first]);
-  }
-}
-
-void send_gossip(AddressKeysetMap& addr_keyset_map, SocketCache& pushers, Serializer* serializer) {
-  unordered_map<string, communication::Request> gossip_map;
-
-  for (auto map_it = addr_keyset_map.begin(); map_it != addr_keyset_map.end(); map_it++) {
-    gossip_map[map_it->first].set_type("PUT");
-
-    for (auto set_it = map_it->second.begin(); set_it != map_it->second.end(); set_it++) {
-      auto res = process_get(*set_it, serializer);
-      if (res.second == 0) {
-        prepare_put_tuple(gossip_map[map_it->first], *set_it, res.first.reveal().value, res.first.reveal().timestamp);
-      }
-    }
-  }
-
-  // send gossip
-  for (auto it = gossip_map.begin(); it != gossip_map.end(); it++) {
-    push_request(it->second, pushers[it->first]);
-  }
-}
 
 // thread entry point
 void run(unsigned thread_id) {
@@ -406,6 +166,7 @@ void run(unsigned thread_id) {
     serializer = new EBSSerializer(thread_id);
   } else {
     logger->info("Invalid node type");
+    exit(1);
   }
 
   // the set of changes made on this thread since the last round of gossip
@@ -474,131 +235,48 @@ void run(unsigned thread_id) {
     // receives a node join
     if (pollitems[0].revents & ZMQ_POLLIN) {
       auto work_start = chrono::system_clock::now();
-      join_handler(SELF_TIER_ID, THREAD_NUM, ip, &join_puller, pushers, 
-                   global_hash_ring_map, local_hash_ring_map,
-                   placement, logger, thread_id, seed, key_stat_map, wt,
-                   join_addr_keyset_map, join_remove_set);
+
+      string message = zmq_util::recv_string(&join_puller);
+      node_join_handler(THREAD_NUM, thread_id, seed, ip, message, logger,
+          global_hash_ring_map, local_hash_ring_map, key_stat_map, placement,
+          join_remove_set, pushers, wt, join_addr_keyset_map);
+
       auto time_elapsed = chrono::duration_cast<chrono::microseconds>(chrono::system_clock::now()-work_start).count();
       working_time += time_elapsed;
       working_time_map[0] += time_elapsed;
     }
 
-    // receive a node departure notice
     if (pollitems[1].revents & ZMQ_POLLIN) {
       auto work_start = chrono::system_clock::now();
+
       string message = zmq_util::recv_string(&depart_puller);
-
-      vector<string> v;
-      split(message, ':', v);
-
-      unsigned tier = stoi(v[0]);
-      string departing_server_ip = v[1];
-      logger->info("Received departure for node {} on tier {}.", departing_server_ip, tier);
-
-      // update hash ring
-      remove_from_hash_ring<GlobalHashRing>(global_hash_ring_map[tier], departing_server_ip, 0);
-
-      if (thread_id == 0) {
-        // tell all worker threads about the node departure
-        for (unsigned tid = 1; tid < THREAD_NUM; tid++) {
-          zmq_util::send_string(message, &pushers[ServerThread(ip, tid).get_node_depart_connect_addr()]);
-        }
-
-        for (auto it = global_hash_ring_map.begin(); it != global_hash_ring_map.end(); it++) {
-          logger->info("hash ring for tier {} size is {}", to_string(it->first), to_string(it->second.size()));
-        }
-      }
+      node_depart_handler(THREAD_NUM, thread_id, ip, message, global_hash_ring_map, logger, pushers);
 
       auto time_elapsed = chrono::duration_cast<chrono::microseconds>(chrono::system_clock::now()-work_start).count();
       working_time += time_elapsed;
       working_time_map[1] += time_elapsed;
     }
 
-    // receives a node departure request
     if (pollitems[2].revents & ZMQ_POLLIN) {
       auto work_start = chrono::system_clock::now();
       string ack_addr = zmq_util::recv_string(&self_depart_puller);
-
-      logger->info("Node is departing.");
-      remove_from_hash_ring<GlobalHashRing>(global_hash_ring_map[SELF_TIER_ID], ip, 0);
-
-      if (thread_id == 0) {
-        for (auto it = global_hash_ring_map.begin(); it != global_hash_ring_map.end(); it++) {
-          auto hash_ring = &(it->second);
-          unordered_set<string> observed_ip;
-
-          for (auto iter = hash_ring->begin(); iter != hash_ring->end(); iter++) {
-            if (observed_ip.find(iter->second.get_ip()) == observed_ip.end()) {
-              zmq_util::send_string(to_string(SELF_TIER_ID) + ":" + ip, &pushers[(iter->second).get_node_depart_connect_addr()]);
-              observed_ip.insert(iter->second.get_ip());
-            }
-          }
-        }
-
-        string msg = "depart:" + to_string(SELF_TIER_ID) + ":" + ip;
-
-        // notify proxies
-        for (auto it = routing_address.begin(); it != routing_address.end(); it++) {
-          zmq_util::send_string(msg, &pushers[RoutingThread(*it, 0).get_notify_connect_addr()]);
-        }
-
-        // notify monitoring nodes
-        for (auto it = monitoring_address.begin(); it != monitoring_address.end(); it++) {
-          zmq_util::send_string(msg, &pushers[MonitoringThread(*it).get_notify_connect_addr()]);
-        }
-
-        // tell all worker threads about the self departure
-        for (unsigned tid = 1; tid < THREAD_NUM; tid++) {
-          zmq_util::send_string(ack_addr, &pushers[ServerThread(ip, tid).get_self_depart_connect_addr()]);
-        }
-      }
-
-      AddressKeysetMap addr_keyset_map;
-      vector<unsigned> tier_ids;
-
-      for (unsigned i = MIN_TIER; i <= MAX_TIER; i++) {
-        tier_ids.push_back(i);
-      }
-
-      bool succeed;
-
-      for (auto it = key_stat_map.begin(); it != key_stat_map.end(); it++) {
-        string key = it->first;
-        auto threads = get_responsible_threads(wt.get_replication_factor_connect_addr(), key, is_metadata(key), global_hash_ring_map, local_hash_ring_map, placement, pushers, tier_ids, succeed, seed);
-
-        if (succeed) {
-          // since we already removed itself from the hash ring, no need to exclude itself from threads
-          for (auto iter = threads.begin(); iter != threads.end(); iter++) {
-            addr_keyset_map[iter->get_gossip_connect_addr()].insert(key);
-          }
-        } else {
-          logger->info("Error: key missing replication factor in node depart routine");
-        }
-      }
-
-      send_gossip(addr_keyset_map, pushers, serializer);
-      zmq_util::send_string(ip + "_" + to_string(SELF_TIER_ID), &pushers[ack_addr]);
+      self_depart_handler(THREAD_NUM, thread_id, seed, ip, ack_addr, logger,
+          global_hash_ring_map, local_hash_ring_map, key_stat_map, placement,
+          routing_address, monitoring_address, wt, pushers, serializer);
 
       auto time_elapsed = chrono::duration_cast<chrono::microseconds>(chrono::system_clock::now()-work_start).count();
       working_time += time_elapsed;
       working_time_map[2] += time_elapsed;
     }
 
-    // receive a request
     if (pollitems[3].revents & ZMQ_POLLIN) {
       auto work_start = chrono::system_clock::now();
-      string serialized_req = zmq_util::recv_string(&request_puller);
-      communication::Request req;
-      req.ParseFromString(serialized_req);
+      string req_string = zmq_util::recv_string(&request_puller);
 
-      // process request
-      auto response = process_request(req, local_changeset, serializer, wt, global_hash_ring_map, local_hash_ring_map, placement, pushers, key_stat_map, key_access_timestamp, total_access, start_time, pending_request_map, seed);
-
-      if (response.tuple_size() > 0 && req.has_respond_address()) {
-        string serialized_response;
-        response.SerializeToString(&serialized_response);
-        zmq_util::send_string(serialized_response, &pushers[req.respond_address()]);
-      }
+      process_user_request(req_string, total_access, seed, start_time,
+          global_hash_ring_map, local_hash_ring_map, key_stat_map,
+          pending_request_map, key_access_timestamp, placement, local_changeset,
+          wt, serializer, pushers);
 
       auto time_elapsed = chrono::duration_cast<chrono::microseconds>(chrono::system_clock::now()-work_start).count();
       working_time += time_elapsed;
@@ -609,13 +287,12 @@ void run(unsigned thread_id) {
     if (pollitems[4].revents & ZMQ_POLLIN) {
       auto work_start = chrono::system_clock::now();
       string serialized_gossip = zmq_util::recv_string(&gossip_puller);
-      communication::Request gossip;
-      gossip.ParseFromString(serialized_gossip);
 
-      //  Process distributed gossip
-      process_gossip(gossip, wt, global_hash_ring_map, local_hash_ring_map, placement, pushers, serializer, key_stat_map, pending_gossip_map, seed);
+      process_gossip(serialized_gossip, wt, global_hash_ring_map, local_hash_ring_map,
+          placement, pushers, serializer, key_stat_map, pending_gossip_map,
+          seed);
+
       auto time_elapsed = chrono::duration_cast<chrono::microseconds>(chrono::system_clock::now()-work_start).count();
-
       working_time += time_elapsed;
       working_time_map[4] += time_elapsed;
     }
@@ -624,158 +301,11 @@ void run(unsigned thread_id) {
     if (pollitems[5].revents & ZMQ_POLLIN) {
       auto work_start = chrono::system_clock::now();
       string serialized_response = zmq_util::recv_string(&replication_factor_puller);
-      communication::Response response;
-      response.ParseFromString(serialized_response);
 
-      vector<string> tokens;
-      split(response.tuple(0).key(), '_', tokens);
-      string key = tokens[1];
-
-      if (response.tuple(0).err_number() == 0) {
-        communication::Replication_Factor rep_data;
-        rep_data.ParseFromString(response.tuple(0).value());
-
-        for (int i = 0; i < rep_data.global_size(); i++) {
-          placement[key].global_replication_map_[rep_data.global(i).tier_id()] = rep_data.global(i).global_replication();
-        }
-
-        for (int i = 0; i < rep_data.local_size(); i++) {
-          placement[key].local_replication_map_[rep_data.local(i).tier_id()] = rep_data.local(i).local_replication();
-        }
-      } else if (response.tuple(0).err_number() == 2) {
-        auto respond_address = wt.get_replication_factor_connect_addr();
-        issue_replication_factor_request(respond_address, key, global_hash_ring_map[1], local_hash_ring_map[1], pushers, seed);
-      } else {
-        for (unsigned i = MIN_TIER; i <= MAX_TIER; i++) {
-          placement[key].global_replication_map_[i] = tier_data_map[i].default_replication_;
-          placement[key].local_replication_map_[i] = DEFAULT_LOCAL_REPLICATION;
-        }
-      }
-
-      if (response.tuple(0).err_number() != 2) {
-        // process pending events
-        vector<unsigned> tier_ids;
-        tier_ids.push_back(SELF_TIER_ID);
-        bool succeed;
-
-        // process pending requests
-        if (pending_request_map.find(key) != pending_request_map.end()) {
-          auto threads = get_responsible_threads(wt.get_replication_factor_connect_addr(), key, is_metadata(key), global_hash_ring_map, local_hash_ring_map, placement, pushers, tier_ids, succeed, seed);
-
-          if (succeed) {
-            bool responsible;
-
-            if (threads.find(wt) != threads.end()) {
-              responsible = true;
-            } else {
-              responsible = false;
-            }
-
-            for (auto it = pending_request_map[key].second.begin(); it != pending_request_map[key].second.end(); it++) {
-              if (!responsible && it->addr_ != "") {
-                communication::Response response;
-
-                if (it->respond_id_ != "") {
-                  response.set_response_id(it->respond_id_);
-                }
-
-                communication::Response_Tuple* tp = response.add_tuple();
-                tp->set_key(key);
-                tp->set_err_number(2);
-
-                for (auto iter = threads.begin(); iter != threads.end(); iter++) {
-                  tp->add_addresses(iter->get_request_pulling_connect_addr());
-                }
-
-                string serialized_response;
-                response.SerializeToString(&serialized_response);
-                zmq_util::send_string(serialized_response, &pushers[it->addr_]);
-              } else if (responsible && it->addr_ == "") {
-                // only put requests should fall into this category
-                if (it->type_ == "P") {
-                  auto current_time = chrono::system_clock::now();
-                  auto ts = generate_timestamp(chrono::duration_cast<chrono::milliseconds>(current_time-start_time).count(), wt.get_tid());
-
-                  process_put(key, ts, it->value_, serializer, key_stat_map);
-                  key_access_timestamp[key].insert(std::chrono::system_clock::now());
-
-                  total_access += 1;
-                  local_changeset.insert(key);
-                } else {
-                  logger->info("Error: GET request with no response address.");
-                }
-              } else if (responsible && it->addr_ != "") {
-                communication::Response response;
-
-                if (it->respond_id_ != "") {
-                  response.set_response_id(it->respond_id_);
-                }
-
-                communication::Response_Tuple* tp = response.add_tuple();
-                tp->set_key(key);
-
-                if (it->type_ == "G") {
-                  auto res = process_get(key, serializer);
-                  tp->set_value(res.first.reveal().value);
-                  tp->set_err_number(res.second);
-
-                  key_access_timestamp[key].insert(std::chrono::system_clock::now());
-                  total_access += 1;
-                } else {
-                  auto current_time = chrono::system_clock::now();
-                  auto ts = generate_timestamp(chrono::duration_cast<chrono::milliseconds>(current_time-start_time).count(), wt.get_tid());
-
-                  process_put(key, ts, it->value_, serializer, key_stat_map);
-                  tp->set_err_number(0);
-
-                  key_access_timestamp[key].insert(std::chrono::system_clock::now());
-                  total_access += 1;
-                  local_changeset.insert(key);
-                }
-
-                string serialized_response;
-                response.SerializeToString(&serialized_response);
-                zmq_util::send_string(serialized_response, &pushers[it->addr_]);
-              }
-            }
-          } else {
-            logger->info("Error: missing key replication factor in process pending request routine.");
-          }
-          pending_request_map.erase(key);
-        }
-
-        // process pending gossip
-        if (pending_gossip_map.find(key) != pending_gossip_map.end()) {
-          auto threads = get_responsible_threads(wt.get_replication_factor_connect_addr(), key, is_metadata(key), global_hash_ring_map, local_hash_ring_map, placement, pushers, tier_ids, succeed, seed);
-
-          if (succeed) {
-            if (threads.find(wt) != threads.end()) {
-              for (auto it = pending_gossip_map[key].second.begin(); it != pending_gossip_map[key].second.end(); it++) {
-                process_put(key, it->ts_, it->value_, serializer, key_stat_map);
-              }
-            } else {
-              unordered_map<string, communication::Request> gossip_map;
-
-              // forward the gossip
-              for (auto it = threads.begin(); it != threads.end(); it++) {
-                gossip_map[it->get_gossip_connect_addr()].set_type("PUT");
-
-                for (auto iter = pending_gossip_map[key].second.begin(); iter != pending_gossip_map[key].second.end(); iter++) {
-                  prepare_put_tuple(gossip_map[it->get_gossip_connect_addr()], key, iter->value_, iter->ts_);
-                }
-              }
-
-              // redirect gossip
-              for (auto it = gossip_map.begin(); it != gossip_map.end(); it++) {
-                push_request(it->second, pushers[it->first]);
-              }
-            }
-          } else {
-            logger->info("Error: missing key replication factor in process pending gossip routine.");
-          }
-          pending_gossip_map.erase(key);
-        }
-      }
+      process_rep_factor_response(serialized_response, seed, total_access,
+          logger, start_time, tier_data_map, global_hash_ring_map, local_hash_ring_map,
+          pending_request_map, pending_gossip_map, key_access_timestamp,
+          placement, key_stat_map, local_changeset, wt, serializer, pushers);
 
       auto time_elapsed = chrono::duration_cast<chrono::microseconds>(chrono::system_clock::now()-work_start).count();
       working_time += time_elapsed;
@@ -785,114 +315,9 @@ void run(unsigned thread_id) {
     // receive replication factor change
     if (pollitems[6].revents & ZMQ_POLLIN) {
       auto work_start = chrono::system_clock::now();
-      logger->info("Received replication factor change.");
       string serialized_req = zmq_util::recv_string(&replication_factor_change_puller);
 
-      if (thread_id == 0) {
-        // tell all worker threads about the replication factor change
-        for (unsigned tid = 1; tid < THREAD_NUM; tid++) {
-          zmq_util::send_string(serialized_req, &pushers[ServerThread(ip, tid).get_replication_factor_change_connect_addr()]);
-        }
-      }
-
-      communication::Replication_Factor_Request req;
-      req.ParseFromString(serialized_req);
-
-      AddressKeysetMap addr_keyset_map;
-      unordered_set<string> remove_set;
-
-      // for every key, update the replication factor and check if the node is still responsible for the key
-      vector<unsigned> tier_ids;
-      for (unsigned i = MIN_TIER; i <= MAX_TIER; i++) {
-        tier_ids.push_back(i);
-      }
-
-      bool succeed;
-
-      for (int i = 0; i < req.tuple_size(); i++) {
-        string key = req.tuple(i).key();
-
-        if (key_stat_map.find(key) != key_stat_map.end()) {
-          auto orig_threads = get_responsible_threads(wt.get_replication_factor_connect_addr(), key, is_metadata(key), global_hash_ring_map, local_hash_ring_map, placement, pushers, tier_ids, succeed, seed);
-
-          if (succeed) {
-            // update the replication factor
-            bool decrement = false;
-            for (int j = 0; j < req.tuple(i).global_size(); j++) {
-              if (req.tuple(i).global(j).global_replication() < placement[key].global_replication_map_[req.tuple(i).global(j).tier_id()]) {
-                decrement = true;
-              }
-
-              placement[key].global_replication_map_[req.tuple(i).global(j).tier_id()] = req.tuple(i).global(j).global_replication();
-            }
-
-            for (int j = 0; j < req.tuple(i).local_size(); j++) {
-              if (req.tuple(i).local(j).local_replication() < placement[key].local_replication_map_[req.tuple(i).local(j).tier_id()]) {
-                decrement = true;
-              }
-
-              placement[key].local_replication_map_[req.tuple(i).local(j).tier_id()] = req.tuple(i).local(j).local_replication();
-            }
-
-            auto threads = get_responsible_threads(wt.get_replication_factor_connect_addr(), key, is_metadata(key), global_hash_ring_map, local_hash_ring_map, placement, pushers, tier_ids, succeed, seed);
-
-            if (succeed) {
-              if (threads.find(wt) == threads.end()) {
-                remove_set.insert(key);
-
-                for (auto it = threads.begin(); it != threads.end(); it++) {
-                  addr_keyset_map[it->get_gossip_connect_addr()].insert(key);
-                }
-              }
-
-              if (!decrement && orig_threads.begin()->get_id() == wt.get_id()) {
-                unordered_set<ServerThread, ThreadHash> new_threads;
-
-                for (auto it = threads.begin(); it != threads.end(); it++) {
-                  if (orig_threads.find(*it) == orig_threads.end()) {
-                    new_threads.insert(*it);
-                  }
-                }
-
-                for (auto it = new_threads.begin(); it != new_threads.end(); it++) {
-                  addr_keyset_map[it->get_gossip_connect_addr()].insert(key);
-                }
-              }
-            } else {
-              logger->info("Error: missing key replication factor in rep factor change routine.");
-            }
-          } else {
-            logger->info("Error: missing key replication factor in rep factor change routine.");
-
-            // just update the replication factor
-            for (int j = 0; j < req.tuple(i).global_size(); j++) {
-              placement[key].global_replication_map_[req.tuple(i).global(j).tier_id()] = req.tuple(i).global(j).global_replication();
-            }
-
-            for (int j = 0; j < req.tuple(i).local_size(); j++) {
-              placement[key].local_replication_map_[req.tuple(i).local(j).tier_id()] = req.tuple(i).local(j).local_replication();
-            }
-          }
-        } else {
-          // just update the replication factor
-          for (int j = 0; j < req.tuple(i).global_size(); j++) {
-            placement[key].global_replication_map_[req.tuple(i).global(j).tier_id()] = req.tuple(i).global(j).global_replication();
-          }
-
-          for (int j = 0; j < req.tuple(i).local_size(); j++) {
-            placement[key].local_replication_map_[req.tuple(i).local(j).tier_id()] = req.tuple(i).local(j).local_replication();
-          }
-        }
-      }
-
-      send_gossip(addr_keyset_map, pushers, serializer);
-
-      // remove keys
-      for (auto it = remove_set.begin(); it != remove_set.end(); it++) {
-        key_stat_map.erase(*it);
-        serializer->remove(*it);
-        local_changeset.erase(*it);
-      }
+      process_rep_factor_change(serialized_req, ip, thread_id, THREAD_NUM, seed, logger, global_hash_ring_map, local_hash_ring_map, placement, key_stat_map, local_changeset, wt, serializer, pushers);
 
       auto time_elapsed = chrono::duration_cast<chrono::microseconds>(chrono::system_clock::now()-work_start).count();
       working_time += time_elapsed;
@@ -1049,7 +474,7 @@ void run(unsigned thread_id) {
       }
     }
 
-    //redistribute data after node joins
+    // redistribute data after node joins
     if (join_addr_keyset_map.size() != 0) {
       unordered_set<string> remove_address_set;
